requirements.txt
altair                    5.4.1
attrs                     25.3.0
blinker                   1.8.2
cachetools                5.5.2
certifi                   2025.1.31
charset-normalizer        3.4.1
click                     8.1.8
contourpy                 1.1.1
cycler                    0.12.1
et_xmlfile                2.0.0
filelock                  3.16.1
fonttools                 4.56.0
fsspec                    2025.3.0
gitdb                     4.0.12
GitPython                 3.1.44
idna                      3.10
importlib_resources       6.4.5
Jinja2                    3.1.6
joblib                    1.4.2
jsonschema                4.23.0
jsonschema-specifications 2023.12.1
kiwisolver                1.4.7
markdown-it-py            3.0.0
MarkupSafe                2.1.5
matplotlib                3.7.5
mdurl                     0.1.2
mpmath                    1.3.0
narwhals                  1.33.0
networkx                  3.1
numpy                     1.24.4
openpyxl                  3.1.5
packaging                 24.2
pandas                    2.0.3
pillow                    10.4.0
pip                       22.1.2
pkgutil_resolve_name      1.3.10
protobuf                  5.29.4
pyarrow                   17.0.0
pydeck                    0.9.1
Pygments                  2.19.1
pyparsing                 3.1.4
python-dateutil           2.9.0.post0
pytz                      2025.2
referencing               0.35.1
requests                  2.32.3
rich                      13.9.4
rpds-py                   0.20.1
scikit-learn              1.3.2
scipy                     1.10.1
setuptools                70.3.0
six                       1.17.0
smmap                     5.0.2
streamlit                 1.40.1
sympy                     1.13.3
tenacity                  9.0.0
threadpoolctl             3.5.0
toml                      0.10.2
torch                     2.4.1
tornado                   6.4.2
typing_extensions         4.13.0
tzdata                    2025.2
urllib3                   2.2.3
wheel                     0.45.1
zipp                      3.20.2
